updatet redaxo and all addons to 8.1

+++ redaxo/src/core/lib/util/file.php
@@ -10,10 +10,34 @@ class rex_file
     /**
      * Returns the content of a file.
      *
-     * @param string $file    Path to the file
-     * @param mixed  $default Default value
+     * @param string $file Path to the file
      *
-     * @return mixed Content of the file or default value if the file isn't readable
+     * @throws rex_exception throws when the file cannot be read
+     *
+     * @return string Content of the file
+     *
+     * @psalm-assert non-empty-string $file
+     */
+    public static function require(string $file): string
+    {
+        return rex_timer::measure(__METHOD__, static function () use ($file) {
+            $content = @file_get_contents($file);
+
+            if (false === $content) {
+                throw new rex_exception('Unable to read file "' . $file . '"');
+            }
+
+            return $content;
+        });
+    }
+
+    /**
+     * Returns the content of a file.
+     *
+     * @template T
+     * @param string $file Path to the file
+     * @param T $default Default value
+     * @return string|T Content of the file or default value if the file isn't readable
      */
     public static function get($file, $default = null)
     {
@@ -26,10 +50,10 @@ class rex_file
     /**
      * Returns the content of a config file.
      *
-     * @param string $file    Path to the file
-     * @param mixed  $default Default value
-     *
-     * @return mixed Content of the file or default value if the file isn't readable
+     * @template T
+     * @param string $file Path to the file
+     * @param T $default Default value
+     * @return array<mixed>|T Content of the file or default value if the file isn't readable
      */
     public static function getConfig($file, $default = [])
     {
@@ -40,10 +64,10 @@ class rex_file
     /**
      * Returns the content of a cache file.
      *
-     * @param string $file    Path to the file
-     * @param mixed  $default Default value
-     *
-     * @return mixed Content of the file or default value if the file isn't readable
+     * @template T
+     * @param string $file Path to the file
+     * @param T $default Default value
+     * @return array<mixed>|T Content of the file or default value if the file isn't readable
      */
     public static function getCache($file, $default = [])
     {
@@ -54,10 +78,12 @@ class rex_file
     /**
      * Puts content in a file.
      *
-     * @param string $file    Path to the file
+     * @param string $file Path to the file
      * @param string $content Content for the file
      *
      * @return bool TRUE on success, FALSE on failure
+     *
+     * @psalm-assert-if-true =non-empty-string $file
      */
     public static function put($file, $content)
     {
@@ -67,8 +93,8 @@ class rex_file
             }
 
             // mimic a atomic write
-            $tmpFile = @tempnam(dirname($file), basename($file));
-            if (false !== file_put_contents($tmpFile, $content) && rename($tmpFile, $file)) {
+            $tmpFile = @tempnam(dirname($file), rex_path::basename($file));
+            if (false !== file_put_contents($tmpFile, $content) && self::move($tmpFile, $file)) {
                 @chmod($file, rex::getFilePerm());
                 return true;
             }
@@ -79,13 +105,51 @@ class rex_file
     }
 
     /**
-     * Puts content in a config file.
+     * Appends content to a file.
      *
-     * @param string $file    Path to the file
-     * @param mixed  $content Content for the file
-     * @param int    $inline  The level where you switch to inline YAML
+     * @param string $file Path to the file
+     * @param string $content Content for the file
+     * @param string $delimiter delimiter for new Content
      *
      * @return bool TRUE on success, FALSE on failure
+     *
+     * @psalm-assert-if-true =non-empty-string $file
+     */
+    public static function append(string $file, string $content, string $delimiter = '')
+    {
+        return rex_timer::measure(__METHOD__, static function () use ($file, $content, $delimiter) {
+            if (!rex_dir::create(dirname($file)) || is_file($file) && !is_writable($file)) {
+                return false;
+            }
+
+            // Check if the file exists and has content
+            $hasContent = is_file($file) && filesize($file) > 0;
+
+            // Append the content to the file with delimiter if it has existing content
+            if ($hasContent) {
+                $content = $delimiter . $content;
+            }
+
+            // Append the content to the file with FILE_APPEND and LOCK_EX flags
+            if (false !== file_put_contents($file, $content, FILE_APPEND | LOCK_EX)) {
+                @chmod($file, rex::getFilePerm());
+                return true;
+            }
+
+            return false;
+        });
+    }
+
+    /**
+     * Puts content in a config file.
+     *
+     * @param string $file Path to the file
+     * @param array<mixed> $content Content for the file
+     * @param int $inline The level where you switch to inline YAML
+     *
+     * @return bool TRUE on success, FALSE on failure
+     *
+     * @psalm-assert-if-true =non-empty-string $file
      */
     public static function putConfig($file, $content, $inline = 3)
     {
@@ -95,10 +159,12 @@ class rex_file
     /**
      * Puts content in a cache file.
      *
-     * @param string $file    Path to the file
-     * @param mixed  $content Content for the file
+     * @param string $file Path to the file
+     * @param array<mixed> $content Content for the file
      *
      * @return bool TRUE on success, FALSE on failure
+     *
+     * @psalm-assert-if-true =non-empty-string $file
      */
     public static function putCache($file, $content)
     {
@@ -112,6 +178,9 @@ class rex_file
      * @param string $dstfile Path of the destination file or directory
      *
      * @return bool TRUE on success, FALSE on failure
+     *
+     * @psalm-assert-if-true =non-empty-string $srcfile
+     * @psalm-assert-if-true =non-empty-string $dstfile
      */
     public static function copy($srcfile, $dstfile)
     {
@@ -119,7 +188,7 @@ class rex_file
             if (is_file($srcfile)) {
                 if (is_dir($dstfile)) {
                     $dstdir = rtrim($dstfile, DIRECTORY_SEPARATOR);
-                    $dstfile = $dstdir . DIRECTORY_SEPARATOR . basename($srcfile);
+                    $dstfile = $dstdir . DIRECTORY_SEPARATOR . rex_path::basename($srcfile);
                 } else {
                     $dstdir = dirname($dstfile);
                     rex_dir::create($dstdir);
@@ -142,10 +211,19 @@ class rex_file
      * @param string $dstfile Path of the destination file or directory
      *
      * @return bool TRUE on success, FALSE on failure
+     *
+     * @psalm-assert-if-true =non-empty-string $srcfile
+     * @psalm-assert-if-true =non-empty-string $dstfile
      */
     public static function move(string $srcfile, string $dstfile): bool
     {
-        return rename($srcfile, $dstfile);
+        if (@rename($srcfile, $dstfile)) {
+            return true;
+        }
+        if (copy($srcfile, $dstfile)) {
+            return unlink($srcfile);
+        }
+        return false;
     }
 
     /**
@@ -158,9 +236,16 @@ class rex_file
     public static function delete($file)
     {
         return rex_timer::measure(__METHOD__, static function () use ($file) {
-            if (is_file($file)) {
-                return unlink($file);
+            $tryUnlink = @unlink($file);
+
+            // re-try without error suppression to compensate possible race conditions
+            if (!$tryUnlink) {
+                clearstatcache(true, $file);
+                if (is_file($file)) {
+                    return unlink($file);
+                }
             }
+
             return true;
         });
     }
@@ -171,6 +256,8 @@ class rex_file
      * @param string $filename Filename
      *
      * @return string Extension of $filename
+     *
+     * @psalm-assert-if-true =non-empty-string $filename
      */
     public static function extension($filename)
     {
@@ -182,7 +269,7 @@ class rex_file
      *
      * @param string $file Path to the file
      *
-     * @return null|string Mime type or `null` if the type could not be detected
+     * @return string|null Mime type or `null` if the type could not be detected
      */
     public static function mimeType($file): ?string
     {
@@ -192,32 +279,28 @@ class rex_file
             return null;
         }
 
-        // map less common types to their more common equivalent
-        static $mapping = [
-            'application/xml' => 'text/xml',
-            'image/svg' => 'image/svg+xml',
-        ];
-
         if ('text/plain' !== $mimeType) {
-            $mimeType = $mapping[$mimeType] ?? $mimeType;
-
-            return $mimeType ?: null;
+            // map less common types to their more common equivalent
+            return match ($mimeType) {
+                'application/xml' => 'text/xml',
+                'image/svg' => 'image/svg+xml',
+                default => $mimeType ?: null,
+            };
         }
 
-        static $extensions = [
+        return match (strtolower(self::extension($file))) {
             'css' => 'text/css',
             'js' => 'application/javascript',
             'svg' => 'image/svg+xml',
-        ];
-
-        return $extensions[strtolower(self::extension($file))] ?? $mimeType;
+            default => $mimeType,
+        };
     }
 
     /**
      * Formates the filesize of the given file into a userfriendly form.
      *
-     * @param string $file   Path to the file
-     * @param array  $format
+     * @param string $file Path to the file
+     * @param array $format
      *
      * @return string Formatted filesize
      */
